Fix unnamed dimension error message.
The thrown exception in VariantDependency prevented the proper
error message from VariantManager to be displayed as a sync
issue during sync (and failed the sync completely.)

This changeset rework the way things happen by relying
ont the presence of a custom syncIssue type (which will
also allow having custom quick fixes). If the sync issue
is created early by the VariantManager, then both the
variantDependency and the ModelBuilder will degrade
to doing less instead of failing due to improper setup.

Change-Id: Iaf62c046421e57a54d5a1c1d90fba0d1de90c3cb
Test: N/A

Former-commit-id: 6a8f789a95ba646a59bd59e32a6ca35a5358b8c8