Postpone enum verification in ResourceValuesXmlParser
Building a symbol table for a library module would fail if we parsed an
enum first before an ID with the same name. Both <enum> and <id>
create "id" type symbols and are allowed to have colliding names.
This CL postpones adding the enum symbols into the symbol table until
we're done parsing all resources in the given directory.

BUG=38413705
Test: new test added in ResourceValuesXmlParserTest
Change-Id: I98f24d54ebdbbc8c0a102974b2e206a3cbfc0aee

Former-commit-id: 2b39a5a1f9963e63db2aa7f79d055a098d949019