Warn about using a dev-mode product flavor
Our docs recommended creating a dev-mode variant
https://developer.android.com/studio/build/optimize-your-build.html#create_dev_variant

This is no longer necessary (except for users who primarily work on
the command line), and has some serious downsides.

This CL locates product flavors called "dev" with minSdkVersion=21
and explains that this is no longer necessary.

It also adds support to the GradleModelMocker for the applicationIdSuffix
and versionNameSuffix properties, which were used in the above dev mode
documentation but was not yet supported by the model mocker.

Test: Unit tests included
Change-Id: I57761c0e85d74cddce616e1ab983d1e0fc75957a

Former-commit-id: fc029b0496d8d0816ef03b96da547013194b85ff