Add support for color spaces in screenshots
RawImage can now make use of version 2 of the screen capture
protocol to read the color space of the screenshot. The
command line screenshot tool is updated to save the color space
as an ICC profile. The utility library pixelprobe can now read
an ICC profile name more accurately (support of Unicode names).

Test: Manual
Bug: 29940137
Change-Id: Iba57fb4d7a341d8efe400ed9e80c78f09562fe24

Former-commit-id: f764d774fbdf4fa7cf258b2e8c328a458a9aca53