Add missing @OutputFiles property in PackageAndroidArtifact.
PackageAndroidArtifact currently specifies the parent directory of
output files as output, but not the output files themselves.

Therefore, when the target ABI or density has changed, some input
properties of PackageAndroidArtifact will change, but not the output
directory. As a result, the subsequent build is considered as
incremental, causing this task to package the *changed* dex files into
a new APK. However, since there could be no changed dex files, the new
APK would then contain no dex files, which leads to failed installation.

This commit adds @OutputFiles property to PackageAndroidArtifact. (A
change in output files will now cause the task to run fully and generate
APKs correctly.)

Bug: 38481325
Test: AbiChangeTest
Change-Id: I521184bd445242170f30c9adf6fa3410be2ea322

Former-commit-id: a1416cb2ecffe5c47750e2b8f13944085f37b144