Fix lint permissions check to support Android Things
Lint permissions check currently require permissions in the manifest as
well as runtime permissions. But in Android Things, we require
permissions only in the manifest. This CL fixes lint to ignore runtime
permission checks if the project is an Android Things project. We check
if it is an Android Things project by checking if the merged manifest
uses the Things library.

Test: added unit tests
Bug: 62904015

Change-Id: Ia09736da6a0a728bb2d61cf5cee32d87bb5122bd

Former-commit-id: 457d28c35c97333db3e60c9a1752ffc7c85b5222