Remove redundant precondition check in ProcessAndroidResources.
A precondition check in ProcessAndroidResources is incorrect since it
assumes buildTargetAbi contains one ABI while in fact it can be
multiple ABIs separated by commas.

ProcessAndroidResources.getSplitsToGenerate() already throws an
exception if no splits can be resolved, so there is no need to have an
additional precondition check (or any logging) when this happens.

Bug: 62909130
Test: New test scenario in InstantRunFullSplitApkTest
Change-Id: I422a9fa19180fa774307d1aa5613302b8fc8ea44

Former-commit-id: 9bc55efc193829f8cf5d0bc3dd3e25cbfb581fe7