Add better error message for disabled new resource processing
Currently we have both new resource processing for libraries and AAPT2
turned on by default. The new resource processing being enabled is a
requirement for using AAPT2, but if someone disables it without
disabling AAPT2 they will get an error that doesn't explain the
situation well:
 ERROR:
 /Users/projectDir/build/intermediates/res/merged/release/anim/abc_fade_in.xml
 uncompiled XML file passed as argument. Must be compiled first into
 .flat file.
  - ERROR:  failed parsing overlays
This CL adds a precondition explaining this requirement better.

Bug: 63133768
Test: N/A
Change-Id: Id23834a23eda29badf864e53dedeaa3f728d129e

Former-commit-id: eb7952a97f939c08807116f8490c7953150ec375