Suppress UnguardedFieldAccess error
runDexer() impl is such that dexing will never be called until
the service is initialised, and that is the single entry point
to the dexInProcess()/dexOutOfProcess() private methods. Hence
the lock is needed only to initialise the static instance of
the dexing service, and it's OK to suppress the unguarded access
inspection within the dexing methods themselves.

(this is part of offline inspections fixing before enabling the
revamped presubmit profile)

Test: N/A
Bug: b/63071706
Change-Id: I3bb4f2ee232dd8131b6e6ab7348f44a565b4d327

Former-commit-id: 383c60ffc652376fe8e44227d71194f086de713a