Change location of resources list file for out of process AAPT2
When calling out of process AAPT2 (currently deprecated, but we might
come back to it) we list resources into a file to significantly
shorten the length of the command (several thousand down to a few
hundred).
If ran for pre-Lollipop we would call AAPT2 in parallel for linking
though and this could result in file locking or wrong reads/writes
to the file. This CL makes the file unique for each call by including
the multi apk configuration in the name and therefore the linking
process thread-safe.

Test: gwt
Change-Id: I6849d33edb456a62fd222f07a5dbe79391f735f0

Former-commit-id: b08ccb547e1cbcada7612e6c6241f081f02e9e6d