Refactor SourceSearcher.
The name didn't reflect what the class did, and the code was quite messy
and difficult to understand or extend.

All of the above problems should be fixed in the new code.

There does exist a further problem: the SourceSearcher class also exists
in the com.android.sdklib.build package. I'm not sure how to go about
using my new version of the code in that package.

Change-Id: I8b9342f928b3097c252a07f1b9d359528bfcb31d
Test: added new tests

Former-commit-id: 0b5adbc85ca040411e9361f3fe8421522c8661c5