Handle windows path names.
The regex for <filename>:<line>:<column> was screwing up on windows
paths, because they have a colon in them. I added a clause to the
regexes that handle this without breaking the POSIX path matching.
I also modified the matching code to be able to test for windows or
POSIX style path names on any platform.

b/63805191

Test: This CL fixes a bug that tests found. I added new tests for the
regexes, that are independent of the error parsing.
Change-Id: I855280dfee50dbae5d22f563f476f58e7f1f1c92

Former-commit-id: 5481a549824392fc18ed085e3d2ed7ee05f092ef