Fixed the way we handle CPU states read from /proc/stat
We're using the values from /proc/stat to obtain the total elapsed CPU
time and one of these values is iowait. However, guest and guest_nice
are already included by user and nice, respectively, so we don't need to
include them in the CPU usage sum.

Also, iowait should be considered as "idle" time, so we removed it from
system usage.

Test: updated existing tests
Bug: 63790627
Bug: 37711005
Change-Id: I745858b03f1fa82d7d3910770c9d1153f26c3086

Former-commit-id: 4b31f041a474c4ade2308d995afdd469546ac3a5