Fix HTTP profiler causing POST exception
Previously, HTTP instrumentation of the "connect" method assumed that,
once connected, we would immediately be interested in the response as
well, so we fetched it and tracked it. (The request was sent off to the
network automatically as a side effect to calling the methods
"getResponseMessage" and "getHeaderFields")

However, according to HttpURLConnection's API, it is valid to call
"connect" without actually sending the request just yet. For example,
the following code is supposed to work:

conn.connect(); // Not required to call explicitly, but allowed
writeDataTo(conn.getOutputStream()); // Modify request body
readDataFrom(conn.getInputStream()); // Send request to network

However, because of our instrumentation, the above code would actually
throw an exception - the call to "getOutputStream" would fail because,
internally, our version of "connect" caused the request to already be
sent out, at which point modifying the request body is not allowed.

We fixed this by postponing calls to tracking the response to as late
as we possibly could (i.e. in "getInputStream" and all methods which,
as a side-effect, would force fetching the response).

Test: N/A
Bug: 63957810
Change-Id: I326a7582a9b72267b9085abcdd37983e4ccc757b

Former-commit-id: 453229ef5fce0395acace2ed24b541fd77161432