Properly substract 2 ArtifactCollections.
The previous implementation depended on ComponentIdentifier being
the same (or different). Hoever ComponentIdentifier does not make
a distinction between 2 dependencies that only differ via a maven
classifier.

In theory ResolvedArtifactResult should contain the classifier in
its getVariant().getAttributeContainer() however that's not the
case right now.

So instead we do just what the FileCollection does and just look
at the artifact file itself.

Bug: 63748587
Change-Id: Icaf6a66e2b325515f25c6b30429dc6788e724b47
Test: TestWithSameDepAsAppWithClassifier

Former-commit-id: c88acc35e1915b74a97ac23d3bd930391fa0b758