Configurable allocation stack depth
Obtaining callstack is the most time consuming code
within the allocaiton callback. To ensure users have
a way to improve performance, this change introduces
a parameter to adjust the depth of callstack to obtain.
The default from Studio is 50 and this can be configured
via the -Dprofiler.alloc.stack.depth property.

Bug: 64122979
Test: N/A
Change-Id: I7313bc466248d7b89c16aa81a29d8e5f985b34c0

Former-commit-id: 60d7e4b3600402367457e3680746f8186d9c7a2c