Fix custom rule loading and testing (62636843 and 37101265)
37101265: Android Studio fails to load custom rules in dependent libs
62636843: CustomRuleTest.testGlobalLintJar is flaky

The first bug fix handles loading all the custom rules that applies to
a project when running in single file analysis (e.g. from Android
Studio).

Adding a regression test for this was tricky, and most of the CL deals
with changing this (e.g. ensuring that model mocking (where I'm
providing a mock AndroidLibrary with a custom lint.jar as a dependency
is initialized early enough before custom rules are registered by the
driver etc.

When adding this I also noticed / remembered that many of the lint
custom rule tests had been disabled because they didn't work under
sharding; tracked down that issue and fixed that as well.

Test: Unit tests updated, including test infrastructure changes to
handle custom rule testing

Change-Id: Ieba940aefef6171e0a2a2fcd914f408047651b1d

Former-commit-id: 892906989f912bc224b7ce099d111853394173de