Disable try-with-resources explicitly for api 19+
Try with resources would be desugared for min
api level 19+, although platform supports it. In
addition to that, runtime classes would not be packaged,
and this led crashes.

This CL passes --nodesugar_try_with_resources_if_needed flag
to desugar if min api is 19+, thus explicitly disabling
try-with-resources desugaring for these api levels.

BUG=64527520
Test: DesugarAppTest

Change-Id: I9d812bb5be452d4f7892c947c3815e94dbd7bc5f

Former-commit-id: 109ed2c65f9cd0f5b4d9118a60904051eab86116