New DSL for matching strategy.
The new DSL is less confusing than the previous
one and is based on the context of what it applies to
instead of being applied directly to the extension.

Fallbacks for build type/flavors:
android {
  buildType {
    qa {
      matchingFallbacks = 'staging', 'debug'
    }
  }
  productFlavors {
    foo {
      matchingFallbacks = 'bar'
    }
  }
}

Missing dimension selection is now also applied either via
defaultConfig, or product Flavor and include both the
selection and its fallbacks.

android {
  defaultConfig {
    missingDimensionStrategy 'color', 'blue'
    missingDimensionStrategy 'shape', 'rect'
  }
  productFlavors {
    debug {
      missingDimensionStrategy 'color', 'indigo', 'blue'
    }
  }
}

Test: Updated FlavorSelectionTest
Change-Id: Ice55e1c4911ff66574735a6eb8552e5c81d51da9

Former-commit-id: 71a216b1307bc9f95aaebf3e9c7f4e202f255f8c