Split the external lib merging to a separate transform.
In the mono dex case, incremental changes would re-merge all the
dependencies which is effectively a performance regression.

With this change, the external dependencies are always merged
before the main merge activity which lower the amount of work
to do in the incremental case where we now merge a lot less
dexes.

Test: added unit tests for new Transform.
Bug=64542310

Change-Id: I55700e6d9a2e4f618208f8d3c8ac9a3be4b8f356

Former-commit-id: 7c62ba4a149b3e9b82734bb430b0418899e09042