[AVD Mgr] Don't disable emulator keyboard input
When new hardware device definitions are available, AVD Manager
updates existing AVDs by applying the definition changes to AVDs'
config.ini files.

The hardware keyboard setting requires special treatment. No current
devices have hardware keyboards, so that device parameter is
false. The emulator, however, uses this parameter not to display a
simulated hardware keyboard, but to enable keyboard input from a
developer's computer keyboard. For this reason, the parameter defaults
to true for new AVDs.

When a hardware definition is updated, we should not apply the keyboard
parameter to existing AVDs.

Bug: 64053647

Test: Added AvdMangerTest::testUpdateDeviceChanged()

Change-Id: Iac88a4de7fead08eaea54812da62a1f776d4a35c

Former-commit-id: 3076b8ea5fed4af966fa4319ca4ec667a065daa7