Add R.java generation with AAPT2 JNI test case to LinkTest
With the recent changes to path handling in AAPT2 it looks like creating
directories based on absolute paths on mac and linux is broken. It was
caught only by the integration tests when I was updating the prebuilts,
should have been caught earlier by the unit tests.
This test will fail with the new AAPT2 prebuilts containing the bug
(currently checked in version doesn't contain this bug, will keep us
from updating until the bug is fixed).

Bug: 62336414
Test: LinkTest
Change-Id: I52e7f1bc49426990b8a544d32eaef2c29bd095d9

Former-commit-id: 0f460a4092adbad5e0e5b9551f2356a05455379c