Basic support for consuming shared libraries.
In the current state, a shared library AAR would contain three copies of
the resources, up from two. Before this is shipped, we want to get back
to two copies, res/ and res.apk by changing AAPT2 to allow the shared
apk to also be consumed as static.

res/

    The non-namespace-aware source resources in the res/ folder For
    consumption by older tools or without using the namespace-aware
    pipeline. This includes the "R.txt" symbol table file, which informs
    the generation of the R class for this library for consuming
    libraries and for the final APK packaging.

res.apk

    The namespace-aware AAPT2 static library For consumption in
    namespace-aware builds.  The AAR also includes an 'R.jar' which
    contains the compiled R class from this static library, which is
    used for compilation only. The final classes bundled in the APK are
    generated by AAPT2 when linking all the static libraries in to the
    final '.ap_' file.

shared.apk

    The signed shared library apk.  Linked against, but not bundled
    inside the app APK.  Must be installed before the APK is launched.
    Re-uses the same R.jar as the AAPT2 static library for compile time,
    and the final R class is provided in the shared library APK at
    runtime.

Prior to this change, the build author could already enable the
namespace-aware resource pipeline by setting

aaptOptions.namespaced = true

Now, by setting the project property (will be DSL later)
android.consumeDependenciesAsSharedLibraries=true and this will mean the
shared.apk is consumed, and the rest of the content of the APK is
treated as provided. The contents of the
SharedLibraryAndroidManifest.xml will be merged in to the consuming
app's manifest.

Not implemented in this change: support for building shared libraries.

Test: NamespacedAarWithSharedLibTest

Change-Id: I55350576eca57faae3be86ae2f383a9bfb590182

Former-commit-id: 2fdb67168464350d85c1015eccd7ad528f19b64c