Add additional lint driver config options: sdk, cache, lint rules, etc
This CL adds a bunch of new XML configuration elements for
the metadata file used to drive lint from for example Bazel.

You can now configure the compileSdkVersion for modules,
the SDK and caching directories to be used, locations
of custom rules to be included, a baseline to apply, etc.

Test: Unit test included
Change-Id: I5979a675228764c73969ccd6d0450b1952020753

Former-commit-id: 878ed83349f536cf4aae0d9c0b574f0092698610