Added I/O events tracker and a caching system for I/O events data
Added I/O Tracker which is called by the interceptors to send data about
file events such as opening, reading from, writing to or closing a file.
The I/O Tracker sends the data to I/O cache to be stored. The caching
system uses a unique id for each session, and saves the data of one session
in the same object. Every reading or writing operation is sent to I/O
speed cache which convert each call to an interval, and reports the
average speed when queried.

Bug: 64666307
Test: go/IoProfilerTesting
Change-Id: I278348ce7c8f426a79e9bec23f9920ae1c20e135

Former-commit-id: 30d3ebee5b8681ce6b8961c7262861c9583bfc91