Fix bazel README.md to describe the proper way to run a single test
The tests inside the android plugin are sharded. Each shard must unzip
the maven offline repo before attempting to run tests. This means that
when we want to run a single test (via test_filter), each shard
unnecessarily spends time doing a whole bunch of setup, and then decides
that there are no tests matching the test_filter.

One easy way to work around this is to specify --test_output=streamed
which disables sharding. So the maven unzip is performed just once.

Test: N/A
Change-Id: I533f744c47bbb2880736cbe28959d61a83b19e0c

Former-commit-id: 080db27076933d105018b6f61befbe48947f3acd