Fix lint.jar intermediate publishing.
We previously published the res folder instead of the lint.jar.
Likely a copy-paste error.
Also, because the lint task reads things from the model rather
than via task output, or dependency collection, and we removed
the generation of the aar, the model pointed to things that
didn't exist. For instance lint.jar was not built/copied, and
the classes.jar of the aar is never built during module to
module dependency.

This change aim to fix this by injecting LintGradleClient with
the actual data (task output or dependency collection), and
by making the task depend on the proper FileCollection instances.

This change also introduce a larger change to have a configuration
to register external custom lint jars.
lintChecks: used to set dependency and resolve locally to
            prepare the packaged lint.jar that is used locally
            and by consumers.
This is global to the module, and there is a single lint.jar for it,
however the publishing is per-variant (since the transitive deps
are per variants), and the same lint.jar is published to all variants.

Test: LintCustomRuleTest is a new integ test that verifies these changes.
Change-Id: I64631d87e5b1cd721a3f6cb414e369f46739a5b1

Former-commit-id: 444817898d78c6b549b95caf601cb83f684e3351