AAPT2 daemon mode: use new-line when calling AAPT2 in daemon mode
Initially the daemon mode in AAPT2 asked for space characters to divide
the arguments, but that would be problematic with file paths containing
space characters. Therefore we switched to a while to a pathSeparator
character, unfortunately it was already used as a separator in the
splits argument to divide the path from the split config. In the end in
this CL we're switching the character for the last time to be the
new-line character. Matching changes were done in AAPT2.

Test: gwt *Split* with daemon mode enabled and newest AAP2
Change-Id: Ic685612dd5dcf12f60361e94b9a525dc046de1d2

Former-commit-id: 14d251b4b4018735e3c70b75faedc76813ad4c94