rm databinding output when running non incremntal.
when MergerResources is running in non incremental mode, it deletes
its output folder but the databinding output folder was not part
of the cleaning.

this created more issues as the mergeResourcesTask would run twice,
once for merging, once for packaging because the
DataBindingBuildInfoTask would depend on the last one registed
which was merge. Removed this task registration madness in the scope
and use proper output dependency.

BUG=64972925
Test: no new test

Change-Id: Ie8f35813ab8f697488c82e7119ea3605f93f25f0

Former-commit-id: a7800d0a9b727d5c0c827d9fa68b2a63039254b9