Remove custom resolution strategy.
This strategy was put in place to reject dependencies that
can't be part of the apps. These includes the broken android.jar
published to maven by third parties as well as open source artifacts
that are part of the Andoid API like httpclient.

This is not working because the rejection cause is not displayed
anywhere so it looks like a broken/unresolved dependencies instead.

We are moving this to a lint error instead.

Bug: 65050576
Test: existing integ tests.
Change-Id: I489ab48f96d1bd0b074ac9062350439aad4c1895
(cherry picked from commit ce391b90ec6b9532d59d696641bb61b979788d1a [formerly 7700459df431d98b57d21efb45ca5ac1bdc26879])

Former-commit-id: ce565af62e15ea9c4a0c2120e9844bb25c5160ad