Ensure no collision in jars of AARs.
In Gradle 4.1, querying for an ArtifactCollection will
not return all the jars of an AAR if they are named the same
(even if they are in different folder). This seems to be an
issue in ComponentFileArtifactIdentifier which does not consider
the full path of the artifact but just the filename.

This should be fixed in Gradle proper but since we need to
ship on 3.0, this works around the problem.

Bug: 65042828
Test: existing ones.
Change-Id: I0658a36aab3683f037fc78a6e6b95b55a619f4d9
(cherry picked from commit 4e673eb9317b166996688d0f6486e06208b9fe48 [formerly c2c90d31fc8d9a649d18657e06df559b3111228d])

Former-commit-id: 81ba6ec1161598755a341754247f9476810ccd69